requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic-settings>=2.0.0
Pillow>=10.0.0
apscheduler>=3.10.0
google-generativeai>=0.8.0
astral>=3.2
python-dotenv>=1.0.0
